+++ NEWS.md
@@ -16,6 +16,8 @@
 
 * When `bs_theme(brand = FALSE)` we now correctly do not apply brand theming when a `_brand.yml` file is present in the project. (#1196)
 
+* Sidebars from `page_sidebar()` and `layout_sidebar()` are now resizable on desktop-sized screen widths, allowing users to resize the sidebar width as desired. (#1217)
+
 # bslib 0.9.0
 
 ## Breaking changes